clean up pin components
